abstract revision proxy model
