+++ database/models/revisions.py
@@ -1,15 +1,5 @@
-from django.db import models
-
-from . import Species
+from . import Species, RevisionProxyMixin
 
 
-class RevisionManager(models.Manager):
-    def get_queryset(self):
-        return super().get_queryset().filter(revision=True)
-
-
-class SpeciesRevision(Species):
-    objects = RevisionManager()
-
-    class Meta:
-        proxy = True
+class SpeciesRevision(Species, RevisionProxyMixin):
+    pass
