+++ setup.cfg
@@ -9,3 +9,4 @@ extend-exclude =
     rmg-models,
 per-file-ignores =
     database/apps.py:F401
+    database/mixins.py:DJ08
